Componentize traceroute button
